+++ servicetalk-concurrent-api/src/main/java/io/servicetalk/concurrent/api/AbstractMergeCompletableOperator.java
@@ -15,16 +15,12 @@
  */
 package io.servicetalk.concurrent.api;
 
-import io.servicetalk.concurrent.Cancellable;
 import io.servicetalk.concurrent.internal.SignalOffloader;
 
-import java.util.concurrent.atomic.AtomicReferenceFieldUpdater;
-import javax.annotation.Nullable;
-
 import static java.util.Objects.requireNonNull;
-import static java.util.concurrent.atomic.AtomicReferenceFieldUpdater.newUpdater;
 
-abstract class AbstractMergeCompletableOperator extends AbstractNoHandleSubscribeCompletable
+abstract class AbstractMergeCompletableOperator<T extends CompletableMergeSubscriber>
+        extends AbstractNoHandleSubscribeCompletable
         implements CompletableOperator {
 
     private final Completable original;
@@ -46,7 +42,7 @@ abstract class AbstractMergeCompletableOperator extends AbstractNoHandleSubscrib
         // side of the asynchronous boundary.
         final Subscriber operatorSubscriber = signalOffloader.offloadSubscriber(
                 contextProvider.wrapCompletableSubscriberAndCancellable(subscriber, contextMap));
-        MergeSubscriber mergeSubscriber = apply(operatorSubscriber);
+        T mergeSubscriber = apply(operatorSubscriber);
         // Subscriber to use to subscribe to the original source. Since this is an asynchronous operator, it may call
         // Cancellable method from EventLoop (if the asynchronous source created/obtained inside this operator uses
         // EventLoop) which may execute blocking code on EventLoop, eg: beforeCancel(). So, we should offload
@@ -60,105 +56,12 @@ abstract class AbstractMergeCompletableOperator extends AbstractNoHandleSubscrib
     }
 
     @Override
-    public abstract MergeSubscriber apply(Subscriber subscriber);
+    public abstract T apply(Subscriber subscriber);
 
     /**
      * Called after the {@code original} {@link Completable} is subscribed and provides a way to subscribe to all other
      * {@link Completable}s that are to be merged with the {@code original} {@link Completable}.
-     * @param subscriber {@link MergeSubscriber} to be used to merge.
+     * @param subscriber {@link T} to be used to merge.
      */
-    abstract void doMerge(MergeSubscriber subscriber);
-
-    abstract static class MergeSubscriber implements Subscriber {
-        private static final AtomicReferenceFieldUpdater<MergeSubscriber, Object> terminalNotificationUpdater =
-                newUpdater(MergeSubscriber.class, Object.class, "terminalNotification");
-        private static final Object ON_COMPLETE = new Object();
-        private static final Object DELIVERED_DELAYED_ERROR = new Object();
-        @SuppressWarnings("unused")
-        @Nullable
-        private volatile Object terminalNotification;
-        private final Subscriber subscriber;
-        private final CancellableStack dynamicCancellable;
-        private final boolean delayError;
-
-        MergeSubscriber(Subscriber subscriber, boolean delayError) {
-            this.subscriber = subscriber;
-            this.delayError = delayError;
-            dynamicCancellable = new CancellableStack();
-            subscriber.onSubscribe(dynamicCancellable);
-        }
-
-        @Override
-        public final void onSubscribe(Cancellable cancellable) {
-            dynamicCancellable.add(cancellable);
-        }
-
-        @Override
-        public final void onComplete() {
-            if (onTerminate()) {
-                tryToCompleteSubscriber();
-            }
-        }
-
-        @Override
-        public final void onError(Throwable t) {
-            for (;;) {
-                Object terminalNotification = this.terminalNotification;
-                if (terminalNotification == null) {
-                    if (terminalNotificationUpdater.compareAndSet(this, null, t)) {
-                        break;
-                    } else if (!delayError) {
-                        // If we are not delaying error notification, and we fail to set the terminal event then that
-                        // means we have already notified the subscriber and should return to avoid duplicate terminal
-                        // notifications.
-                        return;
-                    }
-                } else {
-                    Throwable tmpT = (Throwable) terminalNotification;
-                    tmpT.addSuppressed(t);
-                    t = tmpT;
-                    break;
-                }
-            }
-
-            // If we are delaying the error then we need to prevent concurrent termination with the subscribe() thread,
-            // and we do this with a two phase atomic set to DELIVERED_DELAYED_ERROR.
-            // If we don't delay the error then we never increase the total completion count, so we don't have to worry
-            // about concurrent invocation.
-            if (!delayError || (onTerminate() &&
-                    t == terminalNotificationUpdater.getAndSet(this, DELIVERED_DELAYED_ERROR))) {
-                onError0(t);
-            }
-        }
-
-        final void tryToCompleteSubscriber() {
-            if (terminalNotificationUpdater.compareAndSet(this, null, ON_COMPLETE)) {
-                subscriber.onComplete();
-            } else if (delayError) {
-                // This maybe called from the subscribe() thread and also the Subscriber thread. It is possible the
-                // merge operation already completed successfully in the Subscriber thread, and then the subscribe()
-                // thread observed all merged Completables have completed and so also calls this method.
-                Object maybeThrowable = terminalNotificationUpdater.getAndSet(this, DELIVERED_DELAYED_ERROR);
-                if (maybeThrowable instanceof Throwable) {
-                    onError0((Throwable) maybeThrowable);
-                }
-            }
-        }
-
-        private void onError0(Throwable t) {
-            // If we are delaying error, then everything must have terminated by now, and there is no need to cancel
-            // anything. However if we are not delaying error then we should cancel all outstanding work.
-            if (!delayError) {
-                dynamicCancellable.cancel();
-            }
-            subscriber.onError(t);
-        }
-
-        /**
-         * Called when a terminal event is received.
-         *
-         * @return {@code true} if we are done processing after changing state due to receiving a terminal event.
-         */
-        abstract boolean onTerminate();
-    }
+    abstract void doMerge(T subscriber);
 }
